PR CR - readable test classes
