+++ src/test/java/com/github/lordrex34/config/EnumForConfig.java
@@ -19,12 +19,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.github.lordrex34.config.impl;
+package com.github.lordrex34.config;
 
 /**
  * @author lord_rex
  */
-public interface ITestConfigMarker
+public enum EnumForConfig
 {
-	// marker interface
+	TEST_1,
+	TEST_2,
+	TEST_3;
 }
