+++ src/test/java/com/github/lordrex34/config/TestConfigSet.java
@@ -26,12 +26,13 @@ import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 
 import java.io.IOException;
+import java.util.Set;
 
 import org.junit.Before;
 import org.junit.Test;
 
-import com.github.lordrex34.config.impl.ConfigSetTest;
-import com.github.lordrex34.config.impl.ITestConfigMarker;
+import com.github.lordrex34.config.annotation.ConfigClass;
+import com.github.lordrex34.config.annotation.ConfigField;
 
 /**
  * @author lord_rex
@@ -58,4 +59,36 @@ public class TestConfigSet
 		assertNotNull(ConfigSetTest.TEST_INT_SET);
 		assertFalse(ConfigSetTest.TEST_INT_SET.isEmpty());
 	}
+	
+	@ConfigClass(fileName = "set_test")
+	public static class ConfigSetTest
+	{
+		@ConfigField(name = "TestBooleanSet", value = "true,false,true")
+		public static Set<Boolean> TEST_BOOLEAN_SET;
+		
+		@ConfigField(name = "TestByteSet", value = "1,2,3")
+		public static Set<Byte> TEST_BYTE_SET;
+		
+		@ConfigField(name = "TestShortSet", value = "111,221,443")
+		public static Set<Short> TEST_SHORT_SET;
+		
+		@ConfigField(name = "TestIntSet", value = "1222,4442,9993")
+		public static Set<Integer> TEST_INT_SET;
+		
+		@ConfigField(name = "TestLongSet", value = "12252353252352,443252353253242,993252362673293")
+		public static Set<Long> TEST_LONG_SET;
+		
+		@ConfigField(name = "TestFloatSet", value = "1.,3.2,5.")
+		public static Set<Float> TEST_FLOAT_SET;
+		
+		@ConfigField(name = "TestDoubleSet", value = "4.1,2.3,9.7")
+		public static Set<Double> TEST_DOUBLE_SET;
+		
+		@ConfigField(name = "TestStringSet", value = "This,is,a,string,array,test.")
+		public static Set<String> TEST_STRING_SET;
+		
+		@ConfigField(name = "TestEnumSet", value = "TEST_1,TEST_2")
+		public static Set<EnumForConfig> TEST_ENUM_SET;
+	}
+	
 }
