+++ src/test/java/com/github/lordrex34/config/TestConfigSupplier.java
@@ -26,12 +26,15 @@ import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertThat;
 
 import java.io.IOException;
+import java.lang.reflect.Field;
 
 import org.junit.Before;
 import org.junit.Test;
 
-import com.github.lordrex34.config.impl.ConfigSupplierTest;
-import com.github.lordrex34.config.impl.ITestConfigMarker;
+import com.github.lordrex34.config.annotation.ConfigClass;
+import com.github.lordrex34.config.annotation.ConfigField;
+import com.github.lordrex34.config.lang.ConfigProperties;
+import com.github.lordrex34.config.supplier.IConfigValueSupplier;
 
 /**
  * @author lord_rex
@@ -54,4 +57,37 @@ public class TestConfigSupplier
 		assertThat(ConfigSupplierTest.TEST_STRING_SUPPLIER, is(ConfigSupplierTest.SUPPLIED_STRING_VALUE));
 		assertThat(ConfigSupplierTest.TEST_INTEGER_SUPPLIER, is(ConfigSupplierTest.SUPPLIED_INTEGER_VALUE));
 	}
+	
+	@ConfigClass(fileName = "supplier_test")
+	public static class ConfigSupplierTest
+	{
+		public static final String SUPPLIED_STRING_VALUE = "My name is Lord Rex.";
+		
+		@ConfigField(name = "TestStringSupplier", valueSupplier = MyStringConfigValueSupplier.class, value = "Value is not loaded from here, but from the supplier.")
+		public static String TEST_STRING_SUPPLIER;
+		
+		public static final class MyStringConfigValueSupplier implements IConfigValueSupplier<String>
+		{
+			@Override
+			public String supply(Field field, ConfigField configField, ConfigProperties properties)
+			{
+				return SUPPLIED_STRING_VALUE;
+			}
+		}
+		
+		public static final int SUPPLIED_INTEGER_VALUE = 1234;
+		
+		@ConfigField(name = "TestIntegerSupplier", valueSupplier = MyIntegerConfigValueSupplier.class, value = "Value is not loaded from here, but from the supplier.")
+		public static int TEST_INTEGER_SUPPLIER;
+		
+		public static final class MyIntegerConfigValueSupplier implements IConfigValueSupplier<Integer>
+		{
+			@Override
+			public Integer supply(Field field, ConfigField configField, ConfigProperties properties)
+			{
+				return SUPPLIED_INTEGER_VALUE;
+			}
+		}
+	}
+	
 }
