--- src/test/java/com/github/lordrex34/config/impl/ConfigClassPostLoadHookTest.java
@@ -1,39 +1,0 @@
-/*
- * Copyright (c) 2017 Reginald Ravenhorst <lordrex34@gmail.com>
- * 
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package com.github.lordrex34.config.impl;
-
-import com.github.lordrex34.config.annotation.ConfigClass;
-import com.github.lordrex34.config.annotation.ConfigField;
-import com.github.lordrex34.config.impl.postloadhooks.ConfigClassTestHook;
-
-/**
- * @author lord_rex
- */
-@ConfigClass(fileName = "class_post_load_hook_test", postLoadHook = ConfigClassTestHook.class)
-public class ConfigClassPostLoadHookTest
-{
-	@ConfigField(name = "TestPostInt", value = "129834")
-	public static int TEST_POST_INT;
-	
-	@ConfigField(name = "TestPostString", value = "Any string is good here.")
-	public static String TEST_POST_STRING;
-}
